requirements.txt
Django==1.8.4
dj-database-url==0.3.0
dj-static==0.0.6
django-easy-pjax==1.2.0
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
Pillow==2.3.0
django-widget-tweaks==1.4.1
django-pure-pagination==0.2.1
django-nocaptcha-recaptcha==0.0.18

pyexcel==0.1.7
pyexcel-io==0.0.6
pyexcel-ods==0.0.10
pyexcel-webio==0.0.3
pyexcel-xls==0.0.7
pyexcel-xlsx==0.0.6
django-excel==0.0.2
reportlab==3.2.0
html5lib==0.999
git+https://github.com/bluehawkarthur/pisa.git
django-role-permissions==0.6.2
django-extensions==1.5.9
XlsxWriter==0.7.3
django-imagekit==3.2.7
git+https://github.com/bluehawkarthur/python_stdnum.git
qrcode==5.2.2
git+git://github.com/ojii/pymaging.git#egg=pymaging
git+git://github.com/ojii/pymaging-png.git#egg=pymaging-png
django-preventconcurrentlogins==0.7
